3.4
can work

+++ ColorSensor.h
@@ -3,12 +3,14 @@
 
 #include <Arduino.h>
 
+// 顏色讀取結構
 struct ColorReading {
-  int red;
-  int green;
-  int blue;
+  unsigned long red;
+  unsigned long green;
+  unsigned long blue;
 };
 
+// 顏色感測器函數
 void initColorSensor();
 ColorReading readColor();
